pyproject.toml
[build-system]
requires = ["hatchling>=1.20.0", "hatch-vcs", "hatch-fancy-pypi-readme"]
build-backend = "hatchling.build"

[project]
name = "labelme"
description = "Image Polygonal Annotation with Python"
license = { text = "GPL-3.0-only" }
requires-python = ">=3.9"
version = "5.9.1"
authors = [{ name = "Kentaro Wada", email = "www.kentaro.wada@gmail.com" }]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "imgviz",
  "loguru",
  "matplotlib",
  "natsort>=7.1.0",
  "numpy",
  "osam>=0.2.5",
  "pillow>=2.8",
  "pyqt5>=5.14.0",
  "pyqt5-qt5!=5.15.13 ; sys_platform == 'linux'",
  "pyqt5-qt5!=5.15.11,!=5.15.12,!=5.15.13,!=5.15.14,!=5.15.15,!=5.15.16 ; sys_platform == 'win32'",
  "pyyaml",
  "scikit-image",
  # Tree ring detection dependencies
  "opencv-python>=4.5.0",
  "scipy>=1.7.0",
  "shapely>=1.7.0",
  "torch>=2.0.0",
  "torchvision>=0.15.0",
  "ultralytics>=8.0.0",
  "pandas>=1.3.0",
]

[tool.hatch.metadata.hooks.fancy-pypi-readme]
content-type = "text/markdown"
fragments = [{ path = "README.md" }]

[dependency-groups]
dev = [
  "mypy>=1.15.0",
  "pyqt5-stubs>=5.15.6.0",
  "pytest>=8.3.4",
  "pytest-qt>=4.4.0",
  "ruff~=0.12.11",
  "twine>=6.1.0",
  "types-pillow>=10.2.0.20240822",
  "types-pyyaml>=6.0.12.20241230",
]

[project.scripts]
labelme = "labelme.__main__:main"
labelme_draw_json = "labelme.cli.draw_json:main"
labelme_draw_label_png = "labelme.cli.draw_label_png:main"
labelme_export_json = "labelme.cli.export_json:main"
labelme_on_docker = "labelme.cli.on_docker:main"
labelme_ring_detect = "labelme.cli.ring_detect:main"

[tool.pytest.ini_options]
qt_api = "pyqt5"
markers = [
  "gui: mark a test as a GUI test.",
]

[tool.ruff.lint]
select = [
  "E",  # pycodestyle
  "F",  # pyflakes
  "I",  # isort
  "UP",  # pyupgrade
]

[tool.ruff.lint.isort]
force-single-line = true

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
check_untyped_defs = true
ignore_missing_imports = true
